add VideoAnalysisToolbar component to aggregate all actions outside of the page component
